add transparency func & demo (#59)
1. adjust RGB struct
2. adjust rasterizer
3. add transparency func & demo

+++ platform/BaseInterface.hpp
@@ -21,7 +21,7 @@ class BaseInterface {
 public:
     virtual void drawPixel(uint32_t x, uint32_t y, uint8_t r, uint8_t g, uint8_t b, uint8_t reserved) = 0;
 
-    virtual uint32_t getFrameBufferBase() const = 0;
+    virtual uint64_t getFrameBufferBase() const = 0;
 
     virtual uint32_t getVerticalResolution() const = 0;
